Restructure to MVC with terminal view
