+++ src/view/terminal/config/get.js
@@ -1,4 +1,4 @@
-import { getConfig } from '../../controllers/config-controller.js';
+import { getConfig } from '../../../controllers/config-controller.js';
 
 export function getConfigCommand({ key }) {
   const value = getConfig({ key });
