+++ src/view/terminal/config/set.js
@@ -1,4 +1,4 @@
-import { setConfig } from '../../controllers/config-controller.js';
+import { setConfig } from '../../../controllers/config-controller.js';
 
 export function setConfigCommand({ key, value }) {
   setConfig({ key, value });
