+++ src/view/terminal/index.js
