+++ src/view/terminal/payment/add.js
@@ -1,5 +1,5 @@
-import { addPayment } from '../../controllers/payment-controller.js';
-import { parseDate } from '../../core/date.js';
+import { addPayment } from '../../../controllers/payment-controller.js';
+import { parseDate } from '../../../core/date.js';
 
 export function addPaymentCommand({ client, amount, dueDate }) {
   addPayment({ client, amount: parseFloat(amount), dueDate: parseDate(dueDate) });
