+++ src/view/terminal/payment/link.js
@@ -1,4 +1,4 @@
-import { linkPaymentToTransaction } from '../../controllers/payment-controller.js';
+import { linkPaymentToTransaction } from '../../../controllers/payment-controller.js';
 
 export function linkPaymentCommand({ payment, transaction }) {
   linkPaymentToTransaction({ paymentId: payment, transactionId: transaction });
