+++ src/view/terminal/payment/list.js
@@ -1,4 +1,4 @@
-import { getPayments } from '../../controllers/payment-controller.js';
+import { getPayments } from '../../../controllers/payment-controller.js';
 
 export function listPaymentCommand() {
   const payments = getPayments();
