+++ src/view/terminal/profile/add.js
@@ -1,4 +1,4 @@
-import { addProfile } from '../../controllers/profile-controller.js';
+import { addProfile } from '../../../controllers/profile-controller.js';
 
 export function addProfileCommand({ name }) {
   addProfile({ name });
