+++ src/view/terminal/wallet/adjust.js
@@ -1,4 +1,4 @@
-import { adjustWalletBalance } from '../../controllers/wallet-controller.js';
+import { adjustWalletBalance } from '../../../controllers/wallet-controller.js';
 
 export function adjustBalanceCommand({ wallet, amount }) {
   adjustWalletBalance({ walletId: wallet, amount: parseFloat(amount) });
