+++ src/view/terminal/wallet/set-balance.js
@@ -1,4 +1,4 @@
-import { setWalletBalance } from '../../controllers/wallet-controller.js';
+import { setWalletBalance } from '../../../controllers/wallet-controller.js';
 
 export function setBalanceCommand({ wallet, balance }) {
   setWalletBalance({ walletId: wallet, balance: parseFloat(balance) });
